+++ js/data.js
@@ -25,8 +25,6 @@ const name = [
     'Андрей',
 ];
 
-//Данные для генерации случайного ID объекта
-let genObjId = [1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25]
 
-export {commentsId, avatar, message, name, genObjId}; 
+export {commentsId, avatar, message, name};
